test: 移除 FooContext 类
